requirements.txt
playwright==1.42.0
pytest==8.2.0
allure-pytest==2.13.5
